requirements.txt
Adafruit_GPIO==1.0.3
Adafruit_SSD1306==1.6.2
click==8.1.3
discord.py==2.0.0
discord_ui==5.1.6
minestat==2.3.0
Pillow==9.2.0
python-dotenv==0.20.0
